+++ .eslintrc.js
@@ -1,0 +1,58 @@
+module.exports = {
+    root: true,
+    parser: '@typescript-eslint/parser',
+    parserOptions: {
+        tsconfigRootDir: __dirname,
+        project: ['./tsconfig.json'],
+        createDefaultProgram: true,
+    },
+    plugins: ['@typescript-eslint', 'jest', 'simple-import-sort'],
+    extends: [
+        'eslint:recommended',
+        'plugin:@typescript-eslint/eslint-recommended',
+        'plugin:@typescript-eslint/recommended',
+        'plugin:@typescript-eslint/recommended-requiring-type-checking',
+        'plugin:import/errors',
+        'plugin:import/warnings',
+    ],
+    rules: {
+        'newline-before-return': 'error',
+        '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
+        'simple-import-sort/imports': 'error',
+        'sort-imports': 'off',
+        'import/first': 'error',
+        'import/newline-after-import': 'error',
+        'import/no-duplicates': 'error',
+        'no-multiple-empty-lines': 'error',
+        'lines-between-class-members': 'off',
+        '@typescript-eslint/member-ordering': 'error',
+        '@typescript-eslint/lines-between-class-members': [
+            'error',
+            'always',
+            { exceptAfterSingleLine: true, exceptAfterOverload: true },
+        ],
+        '@typescript-eslint/no-empty-function': 'off',
+        '@typescript-eslint/require-await': 'off',
+        '@typescript-eslint/restrict-template-expressions': 'off',
+    },
+    settings: {
+        'import/parsers': {
+            '@typescript-eslint/parser': ['.ts'],
+        },
+        'import/resolver': {
+            typescript: {
+                alwaysTryTypes: true,
+            },
+        },
+    },
+    overrides: [
+        {
+            files: '*.test.ts',
+            rules: {
+                '@typescript-eslint/no-explicit-any': 'off',
+                '@typescript-eslint/no-unsafe-assignment': 'off',
+                '@typescript-eslint/no-unsafe-member-access': 'off',
+            },
+        },
+    ],
+};