+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- **BREAKING CHANGE** Strategy constructor is not module default export
+
+### Added
+- `confidential` client type, to comply with RFC6749 terminology, `private` client type is deprecated
+- Typescript definitions
+
 
 ## 1.1.0 - 2022-10-21
 ### Added